backend/requirements.txt
fastapi
uvicorn
langchain
langchain-google-genai
python-dotenv
faiss-cpu
langchain-community
sse-starlette